Fix AbstractMLC.getMessageListener Return Type
Getter For MessageListener Tuned With Its Setter

Now that we have removed deprecated setMessageListener,
which used to accept an argument of type 'Object', and now
that we have only one setter which accepts an argument of type
`MessageListener`, it now makes sense to return `MessageListener`
from getter method.

This not only makes getter and setter tuned but would fix the
problem wherein the getter and setter required to be working on
the same parameter type. e.g. in jxm library.