GH-1412: Fix Messaging Template
Resolves https://github.com/spring-projects/spring-amqp/issues/1412

The use of the `defaultDestination` is not correct when the template is
used for both sends and receives because it is a queue name for receives
and a routing key for sends.

Add a new option to use the template's configured default receive queue
for receive only methods. False by default to avoid a breaking change;
it should be true by default in a future release.