GH-2432: Fix Redeclaration of Declarables
Resolves https://github.com/spring-projects/spring-amqp/issues/2432

When a container starts; it looks to see if the context contains a
queue bean for any of its queues and requests the admin to redeclare
the infrastructure.

However, if the queue is declared within a `Declarables`, the check
is not performed and therefore the queue is not declared.

Add a check for `Declarables` beans.

**cherry-pick to 2.4.x**