GH-2439: Fix Bindings with Broker Gen Queue Names
Resolves https://github.com/spring-projects/spring-amqp/issues/2439

To configure a broker-named queue, the `Queue` name is set to an empty
String; the name is later populated when the queue is declared.

However, if such a queue is used in a `BindingBuilder`, the builder accesses
the name before it is populated.

When such a queue is used in a binding, retain a copy of the queue object and
retrieve its actual name when the binding is declared.

Add a unit test for all binding types; also tested with the following, with
both queues being bound properly to the exchange.

```java
@Bean
Queue q1() {
	return QueueBuilder.nonDurable("").build();
}

@Bean
Queue q2() {
	return QueueBuilder.nonDurable("").build();
}

@Bean
Binding b1(Queue q1, DirectExchange ex) {
	return BindingBuilder.bind(q1).to(ex).with("foo");
}

@Bean
Binding b2(Queue q2, DirectExchange ex) {
	return BindingBuilder.bind(q2).to(ex).with("bar");
}

@Bean
DirectExchange ex() {
	return new DirectExchange("ex");
}
```

**cherry-pick to 2.4.x**