GH-2452: Redeclare manual entities automatically (#2453)
Fixes https://github.com/spring-projects/spring-amqp/issues/2452

The queue might be declared manually by an `AmqpAdmin`
and its name simply can be used in the listener container.
When we lose a connection, we would like to have just deleted
manual anonymous queue to be redeclared.

* Introduce `AmqpAdmin.getManualDeclarables()`
* Check for queue name presence in the `manualDeclarables` as well
from a `AbstractMessageListenerContainer.attemptDeclarations()`
* Modify `DirectMessageListenerContainerIntegrationTests.testRecoverDeletedQueueGuts()`
to deal with manual declaration and auto-recovery from the container

**Cherry-pick to `2.4.x`**