Fix repos auth for Dependabot CI only
Apparently regular PR come with `Secrets source: none` (for `on: pull_request` GHA event).
So, no any secrets are available to be propagated down to the repository configurations.
And that is indeed a good sign to not let any unexpected PR to be built.

On the other hand, the Dependabot updates are OK and its PRs come with `Secrets source: dependabot`.
So, we are good to auth repositories to build those PRs and merge.

* Mention secrets limitation in the `pr-build.yml`
* Fix `commercial-settings.gradle` to check if Dependabot is running our project build.
Change repositories respectively, and set provided credentials