GH-2760: Fix endless loop in the `DirectMessageListenerContainer`
Fixes: #2760

When the application starts, a message is sent to another service.
During the start, the Rabbitmq broker crashed and the application began to write endlessly to the log line until it ran out of disk space.
The line was written to the log is `DEBUG(org.springframework.amqp.rabbit.listener.DirectReplyToMessageListenerContainer): Consume from queue amq.rabbitmq.reply-to ignore, container stopping`

**Auto-cherry-pick to `3.1.x`**