Revert back to `registerFeature('optional')`
**Auto-cherry-pick to `3.1.x`**

With a `components.java.addVariantsFromConfiguration(configurations.optional)` and `io.spring.dependency-management`
combination the version is not published for `optional` dependencies.
Does not look like there is an easy way to extract that version somehow.

So, revert back to the `registerFeature('optional')` for time being.
Even if it is reported as deprecated in Gradle