Revert back to `registerFeature('optional')`
With a `components.java.addVariantsFromConfiguration(configurations.optional)` and `io.spring.dependency-management`
combination the version is not published for `optional` dependencies.
Does not look like there is an easy way to extract that version somehow.

So, revert back to the `registerFeature('optional')` for time being.
Even if it is reported as deprecated in Gradle

(cherry picked from commit 2789038f4d69196df57110d664073aadd8b7dcc0)