GH-2917: Add `AbstractConnectionFactory.setAddresses(List<String>)`
Fixes: https://github.com/spring-projects/spring-amqp/issues/2917

In Spring Boot `3.4.0` the `org.springframework.boot.autoconfigure.amqp.RabbitProperties#getAddresses()`
returns a `List<String>` while it previously returned a `String`.
This makes it incompatible with `AbstractConnectionFactory.setAddresses` which accepts a `String`.
Previously we could do `cachingConnectionFactory.setAddresses(rabbitProperties.getAddresses());`,
but now have to go with a `cachingConnectionFactory.setAddresses(String.join(",", rabbitProperties.getAddresses()))`

* Expose `AbstractConnectionFactory.setAddresses(List<String>)` for the mentioned convenience to be able to use `List<String>`
as an alternative to comma-separated string