Attempt to fix race condition in the `CachingConnectionFactory`
There is a logic in the `CachingConnectionFactory.destroy()` where we wait for `inFlightAsyncCloses`.
However, it looks like we first close the main connection with all its channels.
Therefore, we may end up with a `ConcurrentModificationException` when we iterate channels list for clean up,
but in-flight requests still wait for their confirms, therefore they may come back to the cache meanwhile we are trying to clean up