Use `AtomicBoolean` for receive ot ignore other messages
The `RabbitAmqpTemplate.receive()` uses an AMQP 1.0 `Consumer`
to receive a single message.
There is a race condition when we have several messages in the queue,
so after receiving one, this consumer may deliver the next one.

* Use `AtomicBoolean messageReceived` to accept only one message from the consumer.
Since the `CompletableFuture` closes that consumer on its completion, there is just small
chance that a new message would be handed to the handler.
And since we do nothing with this new message, it will come back to the queue after we close consumer.