GH-3055: Deprecate JUnit 4 classes in the project
Fixes: https://github.com/spring-projects/spring-amqp/issues/3055

JUnit 4 is no longer actively maintained, and the last maintenance release was JUnit 4.13.2 in February 2021.
In addition, JUnit 4 has a Java 1.5 baseline.
Furthermore, the JUnit team is planning to release JUnit 6.0 with a Java 17 baseline later this year.

Since JUnit 5 support has been deprecated in Spring Framework, we also follow same theme and
deprecate for removal JUnit 4 rules implementations in favor of respective utils for JUnit 5