GH-3079: Allow ACKs within the ShutdownTimeout period
Fixes: #3079
Issue link: https://github.com/spring-projects/spring-amqp/issues/3079

Remove `cancelled()` condition from the `BlockingQueueConsumer.commitIfNecessary()` to allow initiated acks to be committed.
The cancelled consumer means no new deliveries, but that does not mean we cannot ack outstanding messages.
Actually, this is a leftover after previous fix with assumption that messages have to be returned to the queue with `Recovery` command when consumer is cancelled.

Signed-off-by: Marco Cozzi <marco@codeatlas.it>

**Auto-cherry-pick to `3.2.x` & `3.1.x`**