Fix race condition in the `OptionalPayloadTests`
The test fails sporadically like:
```
java.lang.AssertionError:
Expecting actual:
  ["bar", null, "baz"]
to contain exactly in any order:
  ["foo", null, "bar", "baz"]
but could not find the following elements:
  ["foo"]
```
That means the `listener.latch` is fulfilled, but `ArrayList` state is not deterministic
due to different threads interacting with this collection.

* Wrap `listener.deOptionaled` into a `Collections.synchronizedList()`
and use `synchronized (listener.deOptionaled) {` around assert command.