GH-3122: Log MessageRejectedWhileStoppingException as info in DMLC
Fixed the issue `MessageRejectedWhileStoppingException` should not lead to an error log in all cases

Fixes: https://github.com/spring-projects/spring-amqp/issues/3122

The `MessageRejectedWhileStoppingException` in the `DirectMessageListenerContainer` was always logged at the error level, 
even though it is expected during an orderly shutdown and does not indicate a real error. 
Now, when this exception occurs, it is logged at the info level, reducing unnecessary error noise in logs.

* In `DirectMessageListenerContainer.callExecuteListener()`, exceptions of type `MessageRejectedWhileStoppingException` are now logged at info level. 
All other exceptions continue to be logged as errors.

Signed-off-by: Madan H K <109542799+madanhk18@users.noreply.github.com>