GH-3031: Defer SMLC shutdown for pending replies
This commit introduces a mechanism to delay the shutdown of a
SimpleMessageListenerContainer if there are pending replies for
request/reply operations.

A new functional interface, `PendingReplyProvider`, is introduced and
can be set on the container. `RabbitTemplate` now exposes a
`getPendingReplyCount()` method to serve as this provider.

When the provider is set, the container will wait up to the configured
`shutdownTimeout` for the pending reply count to drop to zero before
proceeding with the consumer cancellation.

Signed-off-by: Jeongjun Min <m3k0813@gmail.com>