GH-3161: Fix `HandlerAdapter` for Kotlin `suspend` functions
Fixes: https://github.com/spring-projects/spring-amqp/issues/3161

The Kotlin `suspend` are async by nature.
However, the `HandlerAdapter` is missing to detect that for its `asyncReplies` property
which is used to set `acknowledgeMode` in the container to the `MANUAL` mode.

* Add `KotlinDetector.isSuspendingFunction()` check to the `HandlerAdapter` for its `asyncReplies` property
* Modify `EnableRabbitKotlinTests` to remove `setAcknowledgeMode(AcknowledgeMode.MANUAL)` from listener container factory beans.
Instead, verify in the respective tests that listener container is switched to the `AcknowledgeMode.MANUAL` due to `true`
outcome from the `KotlinDetector.isSuspendingFunction()` for those `suspend` functions in the test configuration

# Conflicts:
#	spring-rabbit/src/main/java/org/springframework/amqp/rabbit/listener/adapter/HandlerAdapter.java
#	spring-rabbit/src/test/kotlin/org/springframework/amqp/rabbit/annotation/EnableRabbitKotlinTests.kt